Test Accounts
Não deve remover uma conta que tenha uma transação

+++ src/services/accounts.js
@@ -24,7 +24,9 @@ module.exports = (app) => {
         .update(account, '*')
     }
 
-    const remove = (id) => {
+    const remove = async (id) => {
+        const transaction = await app.services.transactions.findOne({acc_id:id})
+        if(transaction) throw new ValidationError('Essa conta possui transações associadas!')
         return app.db('accounts')
         .where({id})
         .delete()
